[mod/#94] service 분리
