+++ data/genre/src/main/java/com/napzak/market/genre/service/PreferredGenreService.kt
@@ -1,0 +1,19 @@
+package com.napzak.market.genre.service
+
+import com.napzak.market.genre.dto.GenreItemsResponse
+import com.napzak.market.remote.model.BaseResponse
+import retrofit2.http.GET
+import retrofit2.http.Query
+
+interface PreferredGenreService {
+    @GET("onboarding/genres")
+    suspend fun getPreferredGenres(
+        @Query("cursor") cursor: Long?,
+    ): BaseResponse<GenreItemsResponse>
+
+    @GET("onboarding/genres/search")
+    suspend fun getPreferredGenreResults(
+        @Query("searchWord") searchWord: String,
+        @Query("cursor") cursor: Long?,
+    ): BaseResponse<GenreItemsResponse>
+}
